Update AnslutaRemote.java
Not sure why you start with 1 here as i did not check all your code yet, but this should at least stop the out of bounds exception. Did not test it yet as i do not have my raspi setup yet. However when packetlength was 0 you still start with 1 which causes the out of bounds exception.

Also from what i can see you start to write in the array with index 1 but you only create the array the exact same size the packetlength is. So even if the packlength is not 0 it will give you an error. Example:
packetlength 5
array.length = 5
array[0] = undefined
array{1] = 1st element
array[2] = 2nd element
...
array{4] = 4th element
The 5th element can not be stored as it is out of bounds already.